Ignore GET_DESCRIPTOR requests for DEBUG type
The lsusb command for newer kernel versions seems to generate a GET_DESCRIPTOR
request for a DEBUG descriptor type.  I'm not sure how to answer that
or if this just means there is something buggy elsewhere.  Returning
an empty packet doesn't seems to generate any damage.

* usb.rs (DescriptorType): Add Debug and a few other types.
(get_descriptor): Send empty replies for Debug.